Project restructure and maven and gradle build file updates to run tests on Maven and to use new test structure in Gradle.
