+++ async-websocket-application/build.gradle
@@ -22,21 +22,14 @@ buildscript {
 // Dependencies
 repositories { mavenCentral(); }
 
-// Add a source set for doing functional tests against a running Liberty
-sourceSets {
-    fvt
-}
-
 dependencies {
     providedCompile group:'javax.websocket', name:'javax.websocket-api', version:'1.1'
     providedCompile group:'javax.enterprise.concurrent', name:'javax.enterprise.concurrent-api', version:'1.0'
     providedCompile group:'javax.json', name:'javax.json-api', version:'1.0'
     providedCompile group:'javax.enterprise', name:'cdi-api', version:'1.2'
     testCompile 'junit:junit:4.12'
-    fvtCompile configurations.testCompile
-    fvtRuntime configurations.testRuntime
-    fvtRuntime 'org.eclipse.jetty.websocket:javax-websocket-client-impl:9.3.2.v20150730'
-    fvtRuntime 'org.glassfish:javax.json:1.0.4'
+    testRuntime 'org.eclipse.jetty.websocket:javax-websocket-client-impl:9.3.2.v20150730'
+    testRuntime 'org.glassfish:javax.json:1.0.4'
 }
 
 sourceCompatibility = 1.7
@@ -76,11 +69,20 @@ publishing {
 task fvt(type: Test, dependsOn: jar) {
     group 'Verification'
     description 'Runs the functional verification tests.'
-    testClassesDir = sourceSets.fvt.output.classesDir
-    classpath = sourceSets.fvt.runtimeClasspath
+    reports.html.destination = file("$buildDir/reports/fvt")
+    reports.junitXml.destination = file("$buildDir/test-results/fvt")
+    include '**/fvt/**'
+    exclude '**/unit/**'
     systemProperties = ['liberty.test.port': getHttpPort()]
     dependsOn publishWar, libertyStart
     finalizedBy libertyStop
+}
+
+test {
+	reports.html.destination = file("$buildDir/reports/unit")
+	reports.junitXml.destination = file("$buildDir/test-results/unit")
+	include '**/unit/**'
+	exclude '**/fvt/**'
 }
 
 String getHttpPort() {
